Updated VS Code Folder to be up-to date
