--- ASCII_Timer_CODE/print_minutes.h
@@ -1,19 +1,0 @@
-/**
- *
- * Solution to course project # 9
- * Introduction to programming course
- * Faculty of Mathematics and Informatics of Sofia University
- * Winter semester 2021/2022
- *
- * @author Filip Filchev
- * @idnumber 0MI0600041
- * @compiler VC
- *
- * File with a Declaration of the Function that sends for Printing the Digits of the Minutes
- *
- */
-#pragma once
-#ifndef PRINT_MINUTES_H
-#define PRINT_MINUTES_H
-void printMinutes(const int minutes, const int lineNumber, const int color1, const int color2);
-#endif
