--- ASCII_Timer_CODE/print_seconds.h
@@ -1,19 +1,0 @@
-/**
- *
- * Solution to course project # 9
- * Introduction to programming course
- * Faculty of Mathematics and Informatics of Sofia University
- * Winter semester 2021/2022
- *
- * @author Filip Filchev
- * @idnumber 0MI0600041
- * @compiler VC
- *
- * File with a Declaration of the Function that sends for Printing the Digits of the Seconds
- *
- */
-#pragma once
-#ifndef PRINT_SECONDS_H
-#define PRINT_SECONDS_H
-void printSeconds(const int seconds, const int lineNumber, int color1, int color2);
-#endif
