+++ ASCII_Timer_CODE/send_time_for_print.h
@@ -9,11 +9,15 @@
  * @idnumber 0MI0600041
  * @compiler VC
  *
- * File with the Declaration of the Function that sends for Printing the Digits of the Hours
+ * File with the Declaration of the Functions that send for Printing the Digits of the Hours, Minutes and Seconds
  *
  */
 #pragma once
-#ifndef PRINT_HOURS_H
-#define PRINT_HOURS_H
+#ifndef TIME_H
+#define TIME_H
+
 void printHours(int hours, int lineNumber, int tensOfNumberCount, bool isHoursLessThanTen, int color1, int color2);
+void printMinutes(const int minutes, const int lineNumber, const int color1, const int color2);
+void printSeconds(const int seconds, const int lineNumber, int color1, int color2);
+
 #endif
